- Arbitrage

+++ book_keeping/order.py
@@ -3,5 +3,5 @@ from dataclasses import dataclass
 
 @dataclass
 class Order:
-    price: float
-    quantity: float
+    price: int
+    quantity: int
